backend/services/python-services/requirements.txt
fastapi
uvicorn[standard]
grpcio
grpcio-tools
pika
ortools
openai
python-dotenv
pydantic
groq

# Development dependencies
pytest
pytest-asyncio
black
isort
ruff
